test: add style utils tests
