+++ packages/ui/src/utils/style/__tests__/boxProps.test.ts
@@ -1,0 +1,23 @@
+import { boxProps } from "../boxProps";
+
+describe("boxProps", () => {
+  it("returns empty classes and style when no props provided", () => {
+    expect(boxProps({})).toEqual({ classes: "", style: {} });
+  });
+
+  it("returns classes for tailwind width/height", () => {
+    const result = boxProps({
+      width: "w-16",
+      height: "h-32",
+      padding: "p-2",
+      margin: "m-3",
+    });
+    expect(result).toEqual({ classes: "w-16 h-32 p-2 m-3", style: {} });
+  });
+
+  it("uses style for numeric or non-tailwind values", () => {
+    const result = boxProps({ width: 200, height: "100%" });
+    expect(result.classes).toBe("");
+    expect(result.style).toEqual({ width: 200, height: "100%" });
+  });
+});
