+++ packages/ui/src/utils/style/__tests__/cn.test.ts
@@ -1,0 +1,11 @@
+import { cn } from "../cn";
+
+describe("cn", () => {
+  it("filters out falsey values", () => {
+    expect(cn("a", "", false, null as unknown as string, undefined, "b")).toBe("a b");
+  });
+
+  it("returns empty string with no arguments", () => {
+    expect(cn()).toBe("");
+  });
+});
